Split up mega-configure function
This was kinda scary, and we can't keep it that way; we're looking at
possibly capturing more info as part of the configure flow to support
package and monorepo configurations.
